Development Files
Development files for the language, written completely in python. The 'run.py' is the main file in the project which helps in running the project. 'main.py' acts as a dependency to the 'run.py' file and 'error_pointer.py' helps in pointing to the syntactical or semantic errors present in the language during runtime.